refactor hot wallet funding
